+++ Com.Adobe.Xmp/Com/adobe/xmp/impl/Latin1Converter.cs
@@ -69,18 +69,18 @@ namespace Com.Adobe.Xmp.Impl
                         case StateStart:
                         default:
                         {
-                            if (b < unchecked((int)(0x7F)))
+                            if (b < unchecked(0x7F))
                             {
                                 @out.Append(unchecked((sbyte)b));
                             }
                             else
                             {
-                                if (b >= unchecked((int)(0xC0)))
+                                if (b >= unchecked(0xC0))
                                 {
                                     // start of UTF8 sequence
                                     expectedBytes = -1;
                                     int test = b;
-                                    for (; expectedBytes < 8 && (test & unchecked((int)(0x80))) == unchecked((int)(0x80)); test = test << 1)
+                                    for (; expectedBytes < 8 && (test & unchecked(0x80)) == unchecked(0x80); test = test << 1)
                                     {
                                         expectedBytes++;
                                     }
@@ -100,7 +100,7 @@ namespace Com.Adobe.Xmp.Impl
 
                         case StateUtf8Char:
                         {
-                            if (expectedBytes > 0 && (b & unchecked((int)(0xC0))) == unchecked((int)(0x80)))
+                            if (expectedBytes > 0 && (b & unchecked(0xC0)) == unchecked(0x80))
                             {
                                 // valid UTF8 char, add to readAheadBuffer
                                 readAheadBuffer[readAhead++] = unchecked((sbyte)b);
@@ -157,14 +157,14 @@ namespace Com.Adobe.Xmp.Impl
         /// <returns>Returns a byte array containing a UTF-8 byte sequence.</returns>
         private static sbyte[] ConvertToUtf8(sbyte ch)
         {
-            int c = ch & unchecked((int)(0xFF));
+            int c = ch & unchecked(0xFF);
             try
             {
-                if (c >= unchecked((int)(0x80)))
+                if (c >= unchecked(0x80))
                 {
-                    if (c == unchecked((int)(0x81)) || c == unchecked((int)(0x8D)) || c == unchecked((int)(0x8F)) || c == unchecked((int)(0x90)) || c == unchecked((int)(0x9D)))
+                    if (c == unchecked(0x81) || c == unchecked(0x8D) || c == unchecked(0x8F) || c == unchecked(0x90) || c == unchecked(0x9D))
                     {
-                        return new sbyte[] { unchecked((int)(0x20)) };
+                        return new sbyte[] { unchecked(0x20) };
                     }
                     // space for undefined
                     // interpret byte as Windows Cp1252 char
